ppl愛用 (1/機率)，毋是log
#537
