+++ 試驗/語言模型/TestKenLM語言模型單元試驗.py
@@ -3,7 +3,6 @@ import itertools
 from math import log10
 import os
 from unittest.case import TestCase
-from unittest.mock import patch
 
 
 from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
@@ -39,15 +38,6 @@ class KenLM語言模型單元試驗(TestCase):
             list(self.媠媠巧靚語言模型.評分(self.媠媠巧靚組物件)),
             [-0.0, log10(1 / 2), log10(1 / 2), -0.0, -0.0],
             self.忍受
-        )
-
-    @patch('臺灣言語工具.語言模型.語言模型.語言模型.評分')
-    def test_媠媠巧靚_perplexity(self, 評分mock):
-        評分mock.return_value = [-0.1, -0.01, -0.4, -0.29]
-        self.assertAlmostEqual(
-            self.媠媠巧靚語言模型.perplexity(self.媠媠巧靚組物件),
-            -0.2,
-            delta=self.忍受
         )
 
     def 陣列比較(self, 結果陣列, 答案陣列, 範圍):
